+++ frontend/quiz/quiz.html
@@ -22,7 +22,7 @@
     <div class="center-header padding-from-top-25">
         <div class="col-sm-4 row-align-flex">
             <div class="col-sm-4"><img ng-src="{{emblem}}" class="half-height-float-left"></div>
-            <div class="col-sm-4"><h2 class="text-to-middle"">TEST</h2></div>
+            <div class="col-sm-4"><h2 class="text-to-middle">TEST</h2></div>
             <div class="col-sm-4"><img ng-src="{{emblem}}" class="half-height-float-right"></div>
         </div>
     </div>
