+++ src/main/java/ee/project/offline/quiz/controller/FileUploadController.java
@@ -1,19 +1,14 @@
 package ee.project.offline.quiz.controller;
 
-import java.io.IOException;
-import java.util.stream.Collectors;
-
 import ee.project.offline.quiz.controller.wrapper.FilePathWrapper;
 import ee.project.offline.quiz.service.storage.StorageService;
-import ee.project.offline.quiz.service.storage.StorageFileNotFoundException;
+import ee.project.offline.quiz.service.storage.exceptions.StorageFileNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @RestController
