+++ src/main/java/ee/project/offline/quiz/service/QuizService.java
@@ -3,9 +3,9 @@ package ee.project.offline.quiz.service;
 import ee.project.offline.quiz.domain.Answer;
 import ee.project.offline.quiz.domain.Question;
 import ee.project.offline.quiz.domain.Quiz;
-import ee.project.offline.quiz.domain.dto.AnswerDTO;
-import ee.project.offline.quiz.domain.dto.QuestionDTO;
-import ee.project.offline.quiz.domain.dto.QuizDTO;
+import ee.project.offline.quiz.domain.dto.quiz.AnswerDTO;
+import ee.project.offline.quiz.domain.dto.quiz.QuestionDTO;
+import ee.project.offline.quiz.domain.dto.quiz.QuizDTO;
 import ee.project.offline.quiz.domain.dto.add.AddAnswerDTO;
 import ee.project.offline.quiz.domain.dto.add.AddQuestionDTO;
 import ee.project.offline.quiz.domain.dto.results.QuestionResultWrapper;
@@ -21,6 +21,8 @@ import ee.project.offline.quiz.repository.QuestionRepository;
 import ee.project.offline.quiz.repository.QuizRepository;
 import ee.project.offline.quiz.repository.log.UserAnswerLogRepository;
 import ee.project.offline.quiz.repository.log.UserQuestionLogRepository;
+import ee.project.offline.quiz.service.exceptions.InvalidQuestionException;
+import ee.project.offline.quiz.service.exceptions.NotEnoughQuestionsException;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
