+++ src/main/java/ee/project/offline/quiz/service/exceptions/InvalidQuestionException.java
@@ -1,4 +1,4 @@
-package ee.project.offline.quiz.service;
+package ee.project.offline.quiz.service.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
