+++ src/main/java/ee/project/offline/quiz/service/storage/exceptions/StorageFileNotFoundException.java
@@ -1,4 +1,4 @@
-package ee.project.offline.quiz.service.storage;
+package ee.project.offline.quiz.service.storage.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
