+++ src/test/java/ee/project/offline/quiz/service/QuizServiceTests.java
@@ -3,21 +3,20 @@ package ee.project.offline.quiz.service;
 import ee.project.offline.quiz.domain.Answer;
 import ee.project.offline.quiz.domain.Question;
 import ee.project.offline.quiz.domain.Quiz;
-import ee.project.offline.quiz.domain.dto.AnswerDTO;
-import ee.project.offline.quiz.domain.dto.QuestionDTO;
-import ee.project.offline.quiz.domain.dto.QuizDTO;
+import ee.project.offline.quiz.domain.dto.quiz.QuizDTO;
 import ee.project.offline.quiz.domain.dto.add.AddAnswerDTO;
 import ee.project.offline.quiz.domain.dto.add.AddQuestionDTO;
 import ee.project.offline.quiz.domain.dto.results.QuestionResultWrapper;
 import ee.project.offline.quiz.domain.dto.results.QuizAnswer;
 import ee.project.offline.quiz.domain.dto.results.QuizResults;
-import ee.project.offline.quiz.domain.log.UserAnswerLog;
 import ee.project.offline.quiz.mapper.QuizMapper;
 import ee.project.offline.quiz.repository.AnswerRepository;
 import ee.project.offline.quiz.repository.QuestionRepository;
 import ee.project.offline.quiz.repository.QuizRepository;
 import ee.project.offline.quiz.repository.log.UserAnswerLogRepository;
 import ee.project.offline.quiz.repository.log.UserQuestionLogRepository;
+import ee.project.offline.quiz.service.exceptions.InvalidQuestionException;
+import ee.project.offline.quiz.service.exceptions.NotEnoughQuestionsException;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
