+++ src/test/java/ee/project/offline/quiz/web/FileUploadTests.java
@@ -1,7 +1,7 @@
 package ee.project.offline.quiz.web;
 
 import ee.project.offline.quiz.service.storage.FileSystemStorageService;
-import ee.project.offline.quiz.service.storage.StorageFileNotFoundException;
+import ee.project.offline.quiz.service.storage.exceptions.StorageFileNotFoundException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +16,6 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.fileUpload;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 //  refrence https://spring.io/guides/gs/uploading-files/
 @RunWith(SpringRunner.class)
